Catalan Word Vectors. Training and analysis
This report shows the study of transforming word vectors or word embeddings. So we create, analyze, and evaluate these word vectors. They are numerical representations about contextualized words into a learned vocabulary. Specifically we are trying to get the Catalan Word Vectors by training the following model.\\
To do that task we are using a database created from Wikipedia articles.