Add files via upload
This report shows the study and development of different models for non-causal language modeling. We are comparing the performance of a simple Feed Forward Neural Network with a Transformer Layer. Therefore we are increasing the number of Transformer Layers of the system and also using Multihead Attention. Finally we have applyed a normalization of Input/Output Shared Embeddings.